Fixes #133
